+++ spec/dummy/app/javascript/packs/application.js
@@ -1,5 +1,6 @@
 import 'crud'
 import 'crud/inputs'
+import 'crud/fontawesome'
 import 'bootstrap'
 
 import './styles/main'