+++ webpacker/fontawesome.js
@@ -1,0 +1,9 @@
+import fontawesome from '@fortawesome/fontawesome'
+import faFilter from '@fortawesome/fontawesome-free-solid/faFilter'
+import faSearch from '@fortawesome/fontawesome-free-solid/faSearch'
+import faSort from '@fortawesome/fontawesome-free-solid/faSort'
+import faSortUp from '@fortawesome/fontawesome-free-solid/faSortUp'
+import faSortDown from '@fortawesome/fontawesome-free-solid/faSortDown'
+import faThList from '@fortawesome/fontawesome-free-solid/faThList'
+
+fontawesome.library.add(faFilter, faSearch, faSort, faSortUp, faSortDown, faThList)